Change stored durations to seconds rather than minutes.
With automatic refresh on the flight record, and values such as fuel
reserve and groundspeed calculated during flight, some values bounce
around because of the granularity of minutes as a divisor.

Note that the command line runner is set to run in production, so
that legacy databases have their durations in minutes converted to
seconds on startup.