Calculate moving average ground speed and fuel rate during flight
Previous approach was to calculate the average fuel rate and speed
for the whole flight but this tends to overstate the fuel rate and
therefore understate the reserve during flight. Using a moving
average provides the most useful estimate of fuel reserve during
cruise, which is the critical time for making decisions about
diverting to an alternate.